MemoRepositoryClass作成
